+++ ReadingOutputFromFLYonPIC_v2/PlottingConfig/TimingDataPlot.py
@@ -1,0 +1,6 @@
+class TimingDataPlot:
+    def __init__(self, caseBasePaths, caseFileNames, caseDataNames, fileName):
+        self.caseBasePaths = caseBasePaths
+        self.caseFileNames = caseFileNames
+        self.caseDataNames = caseDataNames
+        self.fileName = fileName
