--- ReadingOutputFromFLYonPIC_v2/Test.py
@@ -1,1 +1,0 @@
-import ScanSCFLY
